# api_documentation/Core.swagger_collection.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Deep Lynx",
...
    "contact": {}
  },
  "host": "example.com",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/containers": {
      "post": {
        "description": "Creates a new container object. Containers are the root level object and are considered to contain both the ontology(in form of Metatypes, Metatype Keys, and MetatypeRelationships) as well as the data stored under that ontology.\n\nEndpoint will accept both a single container request object, or an array of container request objects",
        "summary": "Create",
        "tags": [
          "Containers"
        ],
        "operationId": "Create",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "description": "Accepts an array of container objects - will attempt to update all of them in a single transaction. If the update fails, none of them will go through.",
        "summary": "Batch Update",
        "tags": [
          "Containers"
        ],
        "operationId": "BatchUpdate",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/BatchUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "List all containers. MUST include the limit/offset query parameters as of 3/26/2020.",
        "summary": "List",
        "tags": [
          "Containers"
        ],
        "operationId": "List",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/containers/{container-id}": {
      "put": {
        "description": "Updates the container. This will fail if a container already exists with the proposed updated name.",
        "summary": "Update",
        "tags": [
          "Containers"
        ],
        "operationId": "Update",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "Fetch container by ID.",
        "summary": "Fetch by ID",
        "tags": [
          "Containers"
        ],
        "operationId": "FetchbyID",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "Archives a Container. This is preferred over deletion as deletion has a cascading effect on the deleted type's keys, relationships, and relationship keys. When in doubt, archive over delete. We'd rather have tombstones than cremating the type.",
        "summary": "Archive",
        "tags": [
          "Containers"
        ],
        "operationId": "Archive",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/containers/{container-id}/data/export": {
      "post": {
        "description": "Create a new data export with the included configuration. Configuration values may be encrypted depending on the adapter you've choosen. See the readme for the exporters for more information.",
        "summary": "New Data Export",
        "tags": [
          "Data Export"
        ],
        "operationId": "NewDataExport",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/NewDataExportRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/containers/{container-id}/data/export/{export-id}": {
      "get": {
        "description": "Fetch a data export record by ID",
        "summary": "Get Data Export",
        "tags": [
          "Data Export"
        ],
        "operationId": "GetDataExport",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "export-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "post": {
        "description": "Start or restart a data export by id.",
        "summary": "Start Data Export",
        "tags": [
          "Data Export"
        ],
        "operationId": "StartDataExport",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "export-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "description": "Stops a data export. Please note that this just sends a \"stop\" signal. The application's export adapter determines how to handle the said signal. In some cases the export stopping might not be immediate.",
        "summary": "Stop Data Export",
        "tags": [
          "Data Export"
        ],
        "operationId": "StopDataExport",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "export-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "example": "{\n\t\"adapter\": \"gremlin\",\n\t\"config\": {\n        \"traversal_source\": \"g\",\n        \"mime_type\": \"application/vnd.gremlin-v2.0+json\",\n        \"graphson_v1\": false,\n        \"user\": \"process.env.GREMLIN_PLUGIN_USER || \"\",\n        \"key\": \"process.env.GREMLIN_PLUGIN_KEY || \"\",\n        \"endpoint\": \"process.env.GREMLIN_PLUGIN_ENDPOINT || localhost\",\n        \"port\": \"process.env.GREMLIN_PLUGIN_PORT || 8182\",\n        \"path\": \"process.env.GREMLIN_PLUGIN_ROOT_PATH || gremlin\"\n    }\n}",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "description": "Deletes a data export record. This does not guarantee the export will stop immediately.",
        "summary": "Delete Data Export",
        "tags": [
          "Data Export"
        ],
        "operationId": "DeleteDataExport",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "container-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "export-id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
    "CreateRequest": {
      "title": "CreateRequest",
      "example": {
        "name": "Test Container",
        "description": "Test Container Description"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "description"
      ]
    },
    "UpdateRequest": {
      "title": "UpdateRequest",
      "example": {
        "description": "Final Test Description"
      },
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        }
      },
      "required": [
        "description"
      ]
    },
    "BatchUpdateRequest": {
      "title": "BatchUpdateRequest",
      "example": {
        "name": "Test Container",
        "description": "Test Container New Description",
        "id": ""
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "description",
        "id"
      ]
    },
...
    "NewDataExportRequest": {
      "title": "NewDataExportRequest",
      "example": {
        "adapter": "gremlin",
        "config": {
          "traversal_source": "g",
          "graphson_v1": false,
          "user": "",
          "key": "",
          "endpoint": "localhost",
          "port": "8182",
          "path": "/gremlin",
          "writes_per_second": 300
        }
      },
      "type": "object",
      "properties": {
        "adapter": {
          "type": "string"
        },
        "config": {
          "$ref": "#/definitions/Config1"
        }
      },
      "required": [
        "adapter",
        "config"
      ]
    },
    "Config1": {
      "title": "Config1",
      "example": {
        "traversal_source": "g",
        "graphson_v1": false,
        "user": "",
        "key": "",
        "endpoint": "localhost",
        "port": "8182",
        "path": "/gremlin",
        "writes_per_second": 300
      },
      "type": "object",
      "properties": {
        "traversal_source": {
          "type": "string"
        },
        "graphson_v1": {
          "type": "boolean"
        },
        "user": {
          "type": "string"
        },
        "key": {
          "type": "string"
        },
        "endpoint": {
          "type": "string"
        },
        "port": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "writes_per_second": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "traversal_source",
        "graphson_v1",
        "user",
        "key",
        "endpoint",
        "port",
        "path",
        "writes_per_second"
      ]
    },
...
  },
...
}